+++ src/main/java/dnd/dnd10_backend/Dnd10BackendApplication.java
@@ -2,7 +2,6 @@ package dnd.dnd10_backend;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
 
 @SpringBootApplication
 public class Dnd10BackendApplication {
