+++ src/main/java/dnd/dnd10_backend/auth/domain/Token.java
@@ -28,7 +28,6 @@ public class Token {
     @Id
     private Long id;
 
-    @Column(length = 300)
     private String refreshToken;
 
     @ManyToOne
