--- src/main/java/dnd/dnd10_backend/calendar/dto/request/CalendarRequestDto.java
@@ -1,15 +1,0 @@
-package dnd.dnd10_backend.calendar.dto.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Getter
-public class CalendarRequestDto {
-
-    private String year;
-
-    private String month;
-}
