+++ src/main/java/dnd/dnd10_backend/config/CorsConfig.java
@@ -1,6 +1,5 @@
 package dnd.dnd10_backend.config;
 
-import dnd.dnd10_backend.config.jwt.JwtProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.cors.CorsConfiguration;
@@ -26,20 +25,11 @@ public class CorsConfig {
         UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
         CorsConfiguration config = new CorsConfiguration();
         config.setAllowCredentials(true);
-        config.addAllowedOrigin("http://localhost:3000");
         config.addAllowedOriginPattern("*");
-
         config.addAllowedHeader("*");
-        config.addAllowedHeader(JwtProperties.AT_HEADER_STRING);
-        config.addAllowedHeader(JwtProperties.RT_HEADER_STRING);
-
         config.addExposedHeader("*");
-        config.addExposedHeader(JwtProperties.AT_HEADER_STRING);
-        config.addExposedHeader(JwtProperties.RT_HEADER_STRING);
-
         config.addAllowedMethod("*");
-
-        source.registerCorsConfiguration("/**", config);
+        source.registerCorsConfiguration("/api/**", config);
         return new CorsFilter(source);
     }
 }
