+++ src/main/java/dnd/dnd10_backend/config/jwt/JwtRequestFilter.java
@@ -5,11 +5,8 @@ import com.auth0.jwt.algorithms.Algorithm;
 import com.auth0.jwt.exceptions.JWTVerificationException;
 import com.auth0.jwt.exceptions.TokenExpiredException;
 import dnd.dnd10_backend.user.repository.UserRepository;
-import dnd.dnd10_backend.user.service.TokenService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
 import org.springframework.web.filter.OncePerRequestFilter;
 
@@ -18,9 +15,6 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-
-import static com.auth0.jwt.JWT.require;
-
 
 /**
  * 패키지명 dnd.dnd10_backend.config.jwt
@@ -33,14 +27,13 @@ import static com.auth0.jwt.JWT.require;
  * [수정내용]
  * 예시) [2022-09-17] 주석추가 - 원지윤
  */
+@RequiredArgsConstructor
 @Component
 public class JwtRequestFilter extends OncePerRequestFilter {
 
-    private final TokenService tokenService;
+    @Autowired
+    UserRepository userRepository;
 
-    public JwtRequestFilter(TokenService tokenService){
-        this.tokenService = tokenService;
-    }
     @Override
     protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
 
@@ -55,10 +48,9 @@ public class JwtRequestFilter extends OncePerRequestFilter {
         Long userCode = null;
 
         try {
-            userCode = require(Algorithm.HMAC512(JwtProperties.SECRET)).build().verify(token)
+            userCode = JWT.require(Algorithm.HMAC512(JwtProperties.SECRET)).build().verify(token)
                     .getClaim("id").asLong();
-            Authentication authentication = tokenService.getAuthentication(userCode);
-            SecurityContextHolder.getContext().setAuthentication(authentication);
+
         } catch (TokenExpiredException e) {
             e.printStackTrace();
             request.setAttribute(JwtProperties.AT_HEADER_STRING, "토큰이 만료되었습니다.");
