+++ src/main/java/dnd/dnd10_backend/user/controller/UserController.java
@@ -1,17 +1,16 @@
 package dnd.dnd10_backend.user.controller;
 
-import dnd.dnd10_backend.common.domain.SingleResponse;
-import dnd.dnd10_backend.common.domain.enums.CodeStatus;
-import dnd.dnd10_backend.common.service.ResponseService;
 import dnd.dnd10_backend.config.jwt.JwtProperties;
-import dnd.dnd10_backend.user.dto.request.UserSaveRequestDto;
-import dnd.dnd10_backend.user.dto.response.UserResponseDto;
+import dnd.dnd10_backend.user.domain.User;
+import dnd.dnd10_backend.user.oauth.domain.OauthToken;
 import dnd.dnd10_backend.user.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
+import java.util.List;
 
 /**
  * 패키지명 dnd.dnd10_backend.user.controller
@@ -23,11 +22,6 @@ import javax.servlet.http.HttpServletRequest;
  * @version 1.0
  * [수정내용]
  * 예시) [2022-09-17] 주석추가 - 원지윤
- * [2023-01-31] 프론트에 토큰 관련 헤더 노출 - 원지윤
- * [2023-02-02] 사용자 정보 조회, 등록 , 수정 api 추가 - 원지윤
- * [2023-02-02] 토큰 관련 컨트롤러 분리 - 원지윤
- * [2023-02-03] 토큰에서 이메일 식별 - 원지윤
- * [2023-02-06] #50 User 엔티티 userResponseDto로 변환 - 이우진
  */
 @RestController
 @RequestMapping("/api")
@@ -36,54 +30,36 @@ public class UserController {
     @Autowired
     private UserService userService;
 
-    @Autowired
-    private ResponseService responseService;
+    // 프론트에서 인가코드 돌려 받는 주소
+    // 인가 코드로 엑세스 토큰 발급 -> 사용자 정보 조회 -> DB 저장 -> jwt 토큰 발급 -> 프론트에 토큰 전달
+    @GetMapping("/oauth/token")
+    public ResponseEntity getLogin(@RequestParam("code") String code) { //(1)
 
-    /**
-     * 사용자 정보 조회
-     * @param
-     * @return
-     */
-    @GetMapping("/user")
-    public ResponseEntity showUser(HttpServletRequest request){
-        String token = request.getHeader(JwtProperties.AT_HEADER_STRING)
-                        .replace(JwtProperties.TOKEN_PREFIX,"");
-        UserResponseDto userResponseDto = new UserResponseDto().of(userService.getUserByEmail(token));
-        SingleResponse<UserResponseDto> response = responseService.getResponse(userResponseDto,
-                                                                CodeStatus.SUCCESS_SEARCHED_USER);
-        return ResponseEntity.ok().body(response);
+        // 넘어온 인가 코드를 통해 access_token 발급
+        OauthToken oauthToken = userService.getAccessToken(code);
+
+        // 발급 받은 accessToken 으로 카카오 회원 정보 DB 저장 후 JWT 를 생성
+        List<String> tokenList = userService.saveUserAndGetToken(oauthToken.getAccess_token());
+
+        HttpHeaders headers = new HttpHeaders();
+        headers.add(JwtProperties.AT_HEADER_STRING, JwtProperties.TOKEN_PREFIX + tokenList.get(0));
+
+        headers.add(JwtProperties.RT_HEADER_STRING, JwtProperties.TOKEN_PREFIX + tokenList.get(1));
+
+        return ResponseEntity.ok().headers(headers).body("success");
     }
 
-    /**
-     * 사용자 등록
-     * @param requestDto
-     * @return
-     */
-    @PostMapping("/user/signup")
-    public ResponseEntity signupUser(@RequestBody UserSaveRequestDto requestDto,
-                                     HttpServletRequest request){
-        String token = request.getHeader(JwtProperties.AT_HEADER_STRING)
-                            .replace(JwtProperties.TOKEN_PREFIX,"");
-        UserResponseDto userResponseDto = userService.saveUser(requestDto, token);
-        SingleResponse<UserResponseDto> response = responseService.getResponse(userResponseDto,
-                                                                CodeStatus.SUCCESS_CREATED_USER);
-        return ResponseEntity.ok().body(response);
-    }
+    @GetMapping("/oauth/token/refresh")
+    public ResponseEntity refresh(HttpServletRequest request){
 
-    /**
-     * 사용자 정보 업데이트
-     * @param requestDto
-     * @return
-     */
-    @PutMapping("/user/update")
-    public ResponseEntity updateUser(@RequestBody UserSaveRequestDto requestDto,
-                                     HttpServletRequest request){
-        String token = request.getHeader(JwtProperties.AT_HEADER_STRING)
-                            .replace(JwtProperties.TOKEN_PREFIX,"");
-        UserResponseDto userResponseDto = userService.saveUser(requestDto, token);
-        SingleResponse<UserResponseDto> response = responseService.getResponse(userResponseDto,
-                                                                    CodeStatus.SUCCESS_UPDATED_USER);
-        return ResponseEntity.ok().body(response);
+        List<String> tokenList = userService.reissueRefreshToken(request.getHeader(JwtProperties.RT_HEADER_STRING).replace(JwtProperties.TOKEN_PREFIX,""));
+
+        HttpHeaders headers = new HttpHeaders();
+        headers.add(JwtProperties.AT_HEADER_STRING, JwtProperties.TOKEN_PREFIX + tokenList.get(0));
+
+        headers.add(JwtProperties.RT_HEADER_STRING, JwtProperties.TOKEN_PREFIX + tokenList.get(1));
+
+        return ResponseEntity.ok().headers(headers).body("success");
     }
 
 }
