+++ src/main/java/dnd/dnd10_backend/user/domain/enums/Role.java
@@ -1,10 +1,5 @@
 package dnd.dnd10_backend.user.domain.enums;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonValue;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
 /**
  * 패키지명 dnd.dnd10_backend.user.domain.enums
  * 클래스명 Role
@@ -16,26 +11,7 @@ import lombok.Getter;
  * [수정내용]
  * 예시) [2022-09-17] 주석추가 - 원지윤
  */
-@Getter
-@AllArgsConstructor
+
 public enum Role {
-    WORKER("WORKER"),
-    MANAGER("MANAGER");
-
-    private final String value;
-
-    @JsonCreator
-    public static Role from(String sub) {
-        for (Role role : Role.values()) {
-            if (role.getValue().equals(sub)) {
-                return role;
-            }
-        }
-        return null;
-    }
-
-    @JsonValue
-    public String getValue() {
-        return value;
-    }
+    WORKER, MANAGER
 }
