+++ src/main/java/dnd/dnd10_backend/user/dto/request/UserRequestDto.java
@@ -1,30 +1,22 @@
 package dnd.dnd10_backend.user.dto.request;
 
 import dnd.dnd10_backend.user.domain.enums.Role;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 /**
  * 패키지명 dnd.dnd10_backend.user.dto.request
- * 클래스명 UserSaveRequestDto
- * 클래스설명 사용자 저장을 위한 request dto
- * 작성일 2023-02-02
+ * 클래스명 UserRequestDto
+ * 클래스설명
+ * 작성일 2023-01-28
  *
  * @author 원지윤
  * @version 1.0
  * [수정내용]
  * 예시) [2022-09-17] 주석추가 - 원지윤
- * [2023-02-03] 이메일 필드 삭제 - 원지윤
  */
 @Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-public class UserSaveRequestDto {
+public class UserRequestDto {
     private Role role;
-    private String phoneNumber;
     private String workPlace;
     private String workTime;
 }
