--- src/main/java/dnd/dnd10_backend/user/service/CustomUserDetailsService.java
@@ -1,31 +1,0 @@
-package dnd.dnd10_backend.user.service;
-
-import dnd.dnd10_backend.user.domain.User;
-import dnd.dnd10_backend.user.repository.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-/**
- * 패키지명 dnd.dnd10_backend.user.service
- * 클래스명 CustomUserDetailsService
- * 클래스설명
- * 작성일 2023-02-03
- *
- * @author 원지윤
- * @version 1.0
- * [수정내용]
- * 예시) [2022-09-17] 주석추가 - 원지윤
- */
-@Service
-public class CustomUserDetailsService implements UserDetailsService {
-    @Autowired
-    private UserRepository userRepository;
-    @Override
-    public UserDetails loadUserByUsername(String kakaoEmail) throws UsernameNotFoundException {
-        User user = userRepository.findByKakaoEmail(kakaoEmail);
-        return user;
-    }
-}
