+++ src/main/java/dnd/dnd10_backend/user/service/UserService.java
@@ -1,16 +1,36 @@
 package dnd.dnd10_backend.user.service;
 
+import com.auth0.jwt.JWT;
+import com.auth0.jwt.JWTVerifier;
 import com.auth0.jwt.algorithms.Algorithm;
-import dnd.dnd10_backend.common.domain.enums.CodeStatus;
-import dnd.dnd10_backend.common.exception.CustomerNotFoundException;
+import com.auth0.jwt.exceptions.JWTVerificationException;
+import com.auth0.jwt.exceptions.TokenExpiredException;
+import com.auth0.jwt.interfaces.DecodedJWT;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import dnd.dnd10_backend.auth.domain.Token;
+import dnd.dnd10_backend.auth.repository.TokenRepository;
 import dnd.dnd10_backend.config.jwt.JwtProperties;
 import dnd.dnd10_backend.user.domain.User;
-import dnd.dnd10_backend.user.dto.request.UserSaveRequestDto;
-import dnd.dnd10_backend.user.dto.response.UserResponseDto;
+import dnd.dnd10_backend.user.oauth.domain.KakaoProfile;
+import dnd.dnd10_backend.user.oauth.domain.OauthToken;
 import dnd.dnd10_backend.user.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.HttpEntity;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpMethod;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.util.LinkedMultiValueMap;
+import org.springframework.util.MultiValueMap;
+import org.springframework.web.client.RestTemplate;
 
+import java.util.ArrayList;
+import java.util.Date;
+import java.util.List;
+
+import static com.auth0.jwt.JWT.create;
 import static com.auth0.jwt.JWT.require;
 
 /**
@@ -23,10 +43,7 @@ import static com.auth0.jwt.JWT.require;
  * @version 1.0
  * [수정내용]
  * 예시) [2022-09-17] 주석추가 - 원지윤
- * [2023-01-21] refresh token에 대한 내용 수정 - 원지윤
- * [2023-02-02] token관련 service 분리 - 원지윤
- * [2023-02-06] getUserByToken 오타 수정 - 이우진
- * [2023-02-06] getUserByEmail 리턴값 엔티티로 수정 - 이우진
+ * [2022-01-21] refresh token에 대한 내용 수정 - 원지윤
  */
 @Service
 public class UserService {
@@ -34,52 +51,227 @@ public class UserService {
     @Autowired
     UserRepository userRepository;
 
+    @Autowired
+    TokenRepository tokenRepository;
+
+    //환경 변수 가져오기
+    @Value("${spring.security.oauth2.client.registration.kakao.client-id}")
+    String client_id;
+
+    @Value("${spring.security.oauth2.client.registration.kakao.client-secret}")
+    String client_secret;
+
+    @Value("${spring.security.oauth2.client.registration.kakao.redirect-uri}")
+    String redirect_uri;
+
     /**
-     * 토큰 정보로 사용자를 조회하는 메소드
-     * @param token access token
-     * @return
+     * 인가 코드를 통해 access_token 발급
+     * @param code 인가 코드
+     * @return access_token
      */
-    public UserResponseDto getUserByToken(final String token){
+    public OauthToken getAccessToken(String code) {
+
+        RestTemplate rt = new RestTemplate();
+
+        HttpHeaders headers = new HttpHeaders();
+        headers.add("Content-type", "application/x-www-form-urlencoded;charset=utf-8");
+
+        // HttpBody 오브젝트 생성
+        MultiValueMap<String, String> params = new LinkedMultiValueMap<>();
+        params.add("grant_type", "authorization_code");
+        params.add("client_id", client_id);
+        params.add("redirect_uri", redirect_uri);
+        params.add("code", code);
+        params.add("client_secret", client_secret);
+
+        HttpEntity<MultiValueMap<String, String>> kakaoTokenRequest =
+                new HttpEntity<>(params, headers);
+
+        ResponseEntity<String> accessTokenResponse = rt.exchange(
+                "https://kauth.kakao.com/oauth/token",
+                HttpMethod.POST,
+                kakaoTokenRequest,
+                String.class
+        );
+
+        ObjectMapper objectMapper = new ObjectMapper();
+        OauthToken oauthToken = null;
+        try {
+            oauthToken = objectMapper.readValue(accessTokenResponse.getBody(), OauthToken.class);
+        } catch (JsonProcessingException e) {
+            e.printStackTrace();
+        }
+
+        return oauthToken;
+    }
+
+    /**
+     * 신규 사용자 저장 및 JWT토큰 생성하는 함수
+     * @param token access_token
+     * @return access token과 refresh token List
+     */
+    public List<String> saveUserAndGetToken(String token) {
+        KakaoProfile profile = findProfile(token);
+        List<String> tokenList = new ArrayList<>();
+        User user = userRepository.findByKakaoEmail(profile.getKakao_account().getEmail());
+        if(user == null) {
+            user = User.builder()
+                    .kakaoId(profile.getId())
+                    .kakaoProfileImg(profile.getKakao_account().getProfile().getProfile_image_url())
+                    .kakaoNickname(profile.getKakao_account().getProfile().getNickname())
+                    .kakaoEmail(profile.getKakao_account().getEmail())
+                    .userRole("ROLE_USER").build();
+
+            userRepository.save(user);
+        }
+        Token refreshToken = tokenRepository.findByUser(user);
+
+        if(refreshToken != null){
+            refreshToken.refreshUpdate(createRefreshToken(user));
+        }
+        else{
+            refreshToken = Token.builder()
+                    .user(user)
+                    .refreshToken(createRefreshToken(user)).build();
+
+        }
+
+        tokenRepository.save(refreshToken);
+        
+        //List에 각각 access token과 refresh token 차례로 넣어줌
+        tokenList.add(createToken(user));
+        tokenList.add(refreshToken.getRefreshToken());
+
+        return tokenList;
+    }
+
+    /**
+     * JWT토큰 생성하는 함수
+     * @param user 사용자
+     * @return 발급한 JWT토큰
+     */
+    public String createToken(User user) {
+
+        String jwtToken = JWT.create()
+
+                .withSubject(user.getKakaoEmail())
+                .withExpiresAt(new Date(System.currentTimeMillis()+ JwtProperties.AT_EXP_TIME))
+
+                .withClaim("id", user.getUserCode())
+                .withClaim("nickname", user.getKakaoNickname())
+
+                .sign(Algorithm.HMAC512(JwtProperties.SECRET));
+        return jwtToken;
+    }
+
+    /**
+     * refresh토큰 생성하는 함수
+     * @param user 사용자
+     * @return 발급한 refresh token
+     */
+    public String createRefreshToken(User user) {
+
+        String refreshToken = JWT.create()
+
+                .withSubject(user.getKakaoEmail())
+                .withExpiresAt(new Date(System.currentTimeMillis()+ JwtProperties.RT_EXP_TIME))
+
+                .withClaim("id", user.getUserCode())
+                .withClaim("nickname", user.getKakaoNickname())
+
+                .sign(Algorithm.HMAC512(JwtProperties.SECRET));
+
+        return refreshToken;
+    }
+
+    /**
+     * refresh token을 받아 access token과 refresh toekn 재발급
+     * @param token refresh token
+     * @return access token과 refresh token List
+     */
+    public List<String> reissueRefreshToken(String token){
+        List<String> tokenList = new ArrayList<>();
+        //refreshToken 만료 확인
+        if(!validateToken(token)){ //refresh token까지 만료 되었을 때
+            throw new RuntimeException("다시 로그인해주세요");
+        }
+        
         Long userCode = require(Algorithm.HMAC512(JwtProperties.SECRET)).build().verify(token)
                 .getClaim("id").asLong();
+
         User user = userRepository.findByUserCode(userCode);
-        if(user == null) throw new CustomerNotFoundException(CodeStatus.NOT_FOUND_USER);
-        return UserResponseDto.of(user);
+        Token refreshToken = tokenRepository.findByUser(user);
+
+        //DB의 refreshToken과 비교
+        if(!refreshToken.getRefreshToken().equals(token)){ //DB의 refresh token과 front로부터 받아온 refresh toekn이 다를 때
+            throw new RuntimeException("옳지 않은 토큰");
+
+        }
+
+        //refresh token 재발급
+        refreshToken.refreshUpdate(createRefreshToken(user));
+        tokenRepository.save(refreshToken);
+        
+        //List에 각각 access token과 refresh token 차례로 넣어줌
+        tokenList.add(createToken(user));
+        tokenList.add(refreshToken.getRefreshToken());
+
+        return tokenList;
+    }
+    
+    /**
+     * 토큰 정보를 검증하는 메서드
+     * @param token 토큰
+     * @return 토큰 검증 여부
+     */
+    public boolean validateToken(String token) {
+        Long userCode = null;
+        try {
+            userCode = require(Algorithm.HMAC512(JwtProperties.SECRET)).build().verify(token)
+                    .getClaim("id").asLong();
+            return true;
+        }catch (TokenExpiredException e) {
+            //토큰이 만료되었습니다.
+            e.printStackTrace();
+        } catch (JWTVerificationException e) {
+            //유효하지 않은 토큰입니다.
+            e.printStackTrace();
+        }
+        return false;
     }
 
     /**
-     * 이메일로 사용자를 조회하는 메소드
+     * 카카오 서버에 접근해서 사용자의 정보를 받아오는 함수
      * @param token access token
-     * @return
+     * @return 사용자의 정보
      */
-    public User getUserByEmail(final String token){
-        String email = require(Algorithm.HMAC512(JwtProperties.SECRET)).build().verify(token)
-                .getClaim("email").asString();
-        //user 찾기
-        User user = userRepository.findByKakaoEmail(email);
-        if(user == null) throw new CustomerNotFoundException(CodeStatus.NOT_FOUND_USER);
+    public KakaoProfile findProfile(String token) {
 
-        return user;
-    }
+        RestTemplate rt = new RestTemplate();
 
-    /**
-     * 사용자 등록하는 메소드
-     * @param requestDto client한테서 받아온 사용자 정보
-     * @return UserResponseDto
-     */
-    public UserResponseDto saveUser(UserSaveRequestDto requestDto, final String token){
-        String email = require(Algorithm.HMAC512(JwtProperties.SECRET)).build().verify(token)
-                .getClaim("email").asString();
-        //user 찾기
-        User user = userRepository.findByKakaoEmail(email);
-        if(user == null) throw new CustomerNotFoundException(CodeStatus.NOT_FOUND_USER);
+        HttpHeaders headers = new HttpHeaders();
+        headers.add("Authorization", "Bearer " + token);
+        headers.add("Content-type", "application/x-www-form-urlencoded;charset=utf-8");
 
-        //requestDto로 user 정보 update
-        user.updateUser(requestDto);
+        HttpEntity<MultiValueMap<String, String>> kakaoProfileRequest =
+                new HttpEntity<>(headers);
 
-        //user 정보 저장
-        userRepository.save(user);
+        // Http 요청 (POST 방식) 후, response 변수에 응답을 받음
+        ResponseEntity<String> kakaoProfileResponse = rt.exchange(
+                "https://kapi.kakao.com/v2/user/me",
+                HttpMethod.POST,
+                kakaoProfileRequest,
+                String.class
+        );
 
-        return UserResponseDto.of(user);
+        ObjectMapper objectMapper = new ObjectMapper();
+        KakaoProfile kakaoProfile = null;
+        try {
+            kakaoProfile = objectMapper.readValue(kakaoProfileResponse.getBody(), KakaoProfile.class);
+        } catch (JsonProcessingException e) {
+            e.printStackTrace();
+        }
+
+        return kakaoProfile;
     }
 }
