+++ src/main/resources/application.yml
@@ -9,6 +9,6 @@ spring:
     driver-class-name: com.mysql.cj.jdbc.Driver
   jpa:
     hibernate:
-      ddl-auto: update
+      ddl-auto: create
     generate-ddl: true
     database-platform: org.hibernate.dialect.MySQL5InnoDBDialect
